+++ packages/cli/src/index.ts
@@ -3,7 +3,7 @@ import { defineCommand } from 'citty';
 export const main = defineCommand({
   meta: {
     name: 'epilot',
-    version: '0.1.144',
+    version: '0.1.146',
     description: 'CLI for epilot APIs',
   },
   args: {